LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigation-back-cross-document-preventDefault-expected.txt


FAIL navigation.back() cross-document cannot be cancelled with the navigate event assert_equals: expected 0 but got 2
